+++ result/Makefile
@@ -4,7 +4,8 @@ CFLAGS = -Wall -Wextra -Werror -g
 MFLAGS = -L ./mlx -lmlx -framework OpenGL -framework AppKit -fsanitize=address
 
 SRCS = ./src/Cub3D.c ./gnl/get_next_line_utils.c ./gnl/get_next_line.c \
-		./src/init_map.c ./src/parse.c ./src/movement.c
+		./src/map_parse.c ./src/parse.c ./src/movement.c ./src/utils.c \
+		./src/init.c ./src/map_setting.c
 OBJS = $(SRCS:.c=.o)
 
 all: $(NAME)
