+++ result/inc/Cub3D.h
@@ -6,7 +6,7 @@
 /*   By: yeju <yeju@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/01 15:28:46 by yeju              #+#    #+#             */
-/*   Updated: 2022/05/05 12:15:20 by yeju             ###   ########.fr       */
+/*   Updated: 2022/05/05 13:16:41 by yeju             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -117,18 +117,21 @@ int key_press(int key, t_info *info);
 int key_release(int key, t_info *info);
 int close_win(int keycode, int x, int y, void *param);
 
+/* utils.c */
+int	utils_white_space(char c);
+void	utils_bzero(void *s, size_t n);
+int	utils_read_another(char *line, int i);
+int	utils_read(char **map, char *line, t_info *info);
 
-/* init_map.c */
-void utils_bzero(void *s, size_t n);
+/* map_parse.c */
 int init_map(t_info *info);
 void get_map_size(t_info *info);
 char **read_world_map(char *argv);
+char *read_line_map(char *argv, t_info info);
 int read_map_path(char *line, char fir, char sec, t_info *info);
 int read_map_color(char *line, char fc, t_info *info);
 int read_map_setting(char *line, int i, t_info *info);
 int read_map(char *line, char **map, int i, int readed);
-int utils_read(char **map, char *line, t_info *info);
-char *read_line_map(char *argv, t_info info);
 
 /* parse.c */
 void set_info_pos(t_info *info);
@@ -137,7 +140,7 @@ void load_texture(t_info *info);
 void set_pos(t_info *info);
 
 /* engine.c */
-int    render(t_info *info);
+int	render(t_info *info);
 
 
 /* utils */
